requirements.txt
alembic==1.13.1
asyncpg==0.29.0
bcrypt==4.0.1
email_validator==2.1.1
fastapi==0.110.1
greenlet==3.0.3
passlib==1.7.4
pyjwt[crypto]==2.8.0
pydantic==2.6.4
pydantic-settings==2.2.1
redis==5.0.3
SQLAlchemy==2.0.29
uvicorn==0.29.0
loguru==0.7.2
httpx==0.27.0
pytest-asyncio==0.23.6
apscheduler==3.10.4
